refactor(web_search): move tools into tools/ subpackage
The web_search env had two tool modules (search.py, scrape.py) sitting
flat alongside env.py plus an empty, half-migrated tools/ directory. Move
them into tools/ to follow the layout convention "one tool/reward -> a
singular module; several -> a plural subpackage with domain-named members"
and to use up the leftover empty dir.

- git mv search.py/scrape.py -> tools/search.py, tools/scrape.py
- add tools/__init__.py re-exporting the toolkits and provider type
- repoint env.py and the two unit-test imports at the new paths
- document the convention in CLAUDE.md (environments/ section + web_search bullet)

No behavior change. ruff, mypy, and web_search unit tests (34) pass.
