+++ src/strands_env/environments/web_search/env.py
@@ -23,8 +23,8 @@ from strands_env.core.environment import Environment, EnvironmentConfig
 from strands_env.core.models import ModelFactory
 from strands_env.core.types import RewardFunction
 
-from .scrape import WebScraperToolkit
-from .search import WebSearchAPIProvider, WebSearchToolkit
+from .tools.scrape import WebScraperToolkit
+from .tools.search import WebSearchAPIProvider, WebSearchToolkit
 
 
 class WebSearchConfig(EnvironmentConfig, total=False):
