+++ tests/unit/environments/web_search/test_scrape.py
@@ -26,7 +26,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 import tiktoken
 
-from strands_env.environments.web_search.scrape import (
+from strands_env.environments.web_search.tools.scrape import (
     RESULT_TEMPLATE,
     SUMMARY_PROMPT_TEMPLATE,
     TOKEN_ENCODING,
