+++ tests/unit/environments/web_search/test_search.py
@@ -14,7 +14,7 @@
 
 """Unit tests for WebSearchToolkit pure-logic methods."""
 
-from strands_env.environments.web_search.search import WebSearchToolkit
+from strands_env.environments.web_search.tools.search import WebSearchToolkit
 
 # ---------------------------------------------------------------------------
 # _apply_blocked_domains
